Add files via upload
Title: Combinatorial privacy: Packing splinters in polytopes at scale for private bit sums via SecureHull
Given the digital nature of mainstream computation, the bit sum is quite a fundamental operation and its applications are in plenty. We walkthrough an example use-case below without loss of generality.Motivating problem of COVID-19 private test statistics:Release of private client level statistics of symptoms such as fever/no-fever, test-results (positive/negative) and so forth for analysis and monitoring by a central server during the COVID-19 pandemic while maintaining the privacy of the clients data is a motivating use-case for example. This enables data-driven decision making at the central authority while not having to forego the critical privacy constraints and legal regulations of the clients data.

Abstract:
We present a scheme to obtain counts of 0's and 1's at a server based on private bit streams hosted by multiple clients. The goal is to obtain this solution at the server while maintaining privacy of client data. The bit sums need to be obtained with respect to data from all clients; and not at a per client granularity.
In our scheme called 'SecureHull', we hide the private data encoded as permutations amidst publicly shareable permutation matrices and form a secret doubly stochastic matrix via a convex combination with secret coefficients. We exploit the non-uniqueness of the Birkhoff-von Neumann decomposition and use some remnants of the splintering scheme to provide an unconventional secure computation method to this private bitsum problem. This scheme does not require any private data-dependent communication with the server as is ideal. We also provide lower bounds to quantify the probability of a successful attack. We show that the lower bound can be quadratically reduced with a linear increase in communication up to a constant. Our solution also involves a cryptographic shuffling routine that scales linearly with number of clients as against to the size of the datasets. The rest of the operations do not require a cryptographic approach and are secured through our scheme thereby benefiting its scalability.